+++ src/__init__.py
@@ -1,6 +1,7 @@
 from constantes import *
 from joueur import *
+from bonus import *
 from projectile import *
 from ennemis import *
+from vague import *
 from gamebar import *
-from bonus import *
