+++ src/bonus.py
@@ -1,20 +1,23 @@
 import pygame
+from constantes import *
 
 class Bonus(pygame.sprite.Sprite):
-    """classe représentant un projectile
+    """classe représentant un bonus
     """
-    def __init__(self, coord, vitesse, type):
+    def __init__(self, coord, type):
         pygame.sprite.Sprite.__init__(self)
         
-        match type:
-            case 1:
+        self.type = type
+        match self.type:
+            case 'munitions':
                 self.image = pygame.image.load("images/autres/munitions.png").convert_alpha()
-            case 2:
+            case 'explosif':
                 self.image = pygame.image.load("images/autres/explosif.png").convert_alpha()
+
         self.rect = self.image.get_rect()
         self.rect.midbottom = coord
         
-        self.vitesse = vitesse
+        self.vitesse = vitesse_bonus
 
     def update(self):
         self.rect.top += self.vitesse
