+++ src/constantes.py
@@ -7,7 +7,7 @@ vitesse_joueur = 5
 vitesse_projectile_joueur = 10
 chargeur_joueur = 30
 
-vitesse_bonus = 1
+vitesse_bonus = 2
 
 vie_petit = 2
 vitesse_petit = 3
