+++ src/gamebar.py
@@ -15,6 +15,7 @@ class gameBar(pygame.Surface):
         self.demi = pygame.image.load("images/autres/demi_coeur.png").convert_alpha()
         
         self.munition = pygame.image.load("images/autres/logo_munitions.png").convert_alpha()
+        self.explosif = pygame.image.load("images/autres/logo_explosif.png").convert_alpha()
         
     def draw(self, surface):
         surface.blit(self, (0, 0))
@@ -28,11 +29,20 @@ class gameBar(pygame.Surface):
             surface.blit(self.demi, ((self.joueur.vie // 2)*self.coeur.get_width(), 0))
             
             
-        #affichage des informations sur les projectiles
+        #affichage des informations sur le chargeur
         nb_munitions = self.font.render(str(self.joueur.chargeur), True, "WHITE")
-        pos = self.get_width() - 10 - nb_munitions.get_width()
+        pos = self.get_width() - 20 - nb_munitions.get_width()
         surface.blit(nb_munitions, (pos, self.get_height() // 2 - nb_munitions.get_height() // 2))
         
         pos -= self.munition.get_width() + 20
         surface.blit(self.munition, (pos, self.get_height() // 2 - self.munition.get_height() // 2))
+        
+        #affichage des informations sur les explosifs
+        if self.joueur.explosifs:
+            pos -= self.munition.get_width() + 60
+            nb_explosifs = self.font.render(str(self.joueur.explosifs), True, "WHITE")
+            surface.blit(nb_explosifs, (pos, self.get_height() // 2 - nb_explosifs.get_height() // 2))
+            
+            pos -= self.explosif.get_width() + 20
+            surface.blit(self.explosif, (pos, self.get_height() // 2 - self.explosif.get_height() // 2))
